Move log module into utility package
