+++ tests/utility/__init__.py
@@ -1,0 +1,1 @@
+"""Utility package."""
